Add TIM3 for scanning the RPM of DC motor
Add timer integration for measuring rpm of dc motor. This is the first stage of implementation,
when only the first HAL sensor is supported and connected to STM. In the second phase another
HAL sensor will be connected to TIM4.
The USART2 is used to display measured frequency of DC Motor.